Migrate terraform execution to terraform-exec
Move exec tests to exec_mock_test.go since they don't use the concrete
implementation.
